Add salary changes report wizard
